🎄Implement Instance picker #1
I actually wrote the code in this commit a long time ago. I was not able to check and push it due to some issues. Guess what? The service I used in InstancePicker is now dead. I will keep this simple implementation and mock the instance list data for now using the provided query.